geode: fix compile on x86-64 in tinderbox.
geode_ddc.c: In function 'GeodeI2CInit':
geode_ddc.c:158: error: cast to pointer from integer of different size

+++ src/geode_ddc.c
@@ -155,7 +155,7 @@ GeodeI2CInit(ScrnInfoPtr pScrni, I2CBusPtr * ptr, char *name)
 
     bus->I2CGetBits = geode_ddc_getbits;
     bus->I2CPutBits = geode_ddc_putbits;
-    bus->DriverPrivate.ptr = (void *)(ddciobase);
+    bus->DriverPrivate.ptr = (void *)(unsigned long)(ddciobase);
 
     if (!xf86I2CBusInit(bus))
 	return FALSE;
